Use %run instead of %load to fix compatibility with ipython 5. Fixes #13
This is because in ipython 5, %load sets the cursor at the beginning of
the loaded code and you have to scroll through it and hit enter at the
end to execute. We do not have a reliable way of doing that.

Using %run is kind of suboptimal because it doesn’t show the executed
code in ipython history anymore, but at least it works.